Refs #37159 - Refactor content registration to record discovery class
