+++ lib/katello/repository_types/yum.rb
@@ -1,6 +1,7 @@
 Katello::RepositoryTypeManager.register(::Katello::Repository::YUM_TYPE) do
   pulp3_service_class Katello::Pulp3::Repository::Yum
   pulp3_api_class Katello::Pulp3::Api::Yum
+  repo_discovery_class ::Katello::Resources::Discovery::Yum
   pulp3_plugin 'rpm'
   prevent_unneeded_metadata_publish
