+++ test/lib/repo_discovery_test.rb
@@ -13,7 +13,7 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      rd = RepoDiscovery.create_for('yum').new(base_url, nil, nil, crawled, found, to_follow)
+      rd = RepoDiscovery.class_for('yum').new(base_url, crawled, found, to_follow)
 
       rd.run(to_follow.shift)
       assert_equal 1, rd.crawled.size
@@ -27,17 +27,17 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      rd = RepoDiscovery.new(base_url, 'yum', nil, nil, crawled, found, to_follow)
+      rd = RepoDiscovery.class_for('yum').new(base_url, crawled, found, to_follow)
       rd.stubs(:proxy).returns(@http_proxy)
 
       expected_proxy_params = {
-        proxy_host: @proxy_uri.host,
-        proxy_port: @proxy_uri.port,
-        proxy_user: @http_proxy.username,
-        proxy_password: @http_proxy.password
+        host: "mytest.com",
+        port: 443,
+        user: nil,
+        password: nil
       }
 
-      Anemone.expects(:crawl).with(rd.uri(base_url), expected_proxy_params).returns
+      Spidr.expects(:site).with(base_url, proxy: expected_proxy_params).returns
 
       rd.run(to_follow.shift)
     end
@@ -47,8 +47,8 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      rd = RepoDiscovery.new(base_url, 'yum', nil, nil, crawled, found, to_follow)
-      Anemone.expects(:crawl).with(rd.uri(base_url), {}).returns
+      rd = RepoDiscovery.class_for('yum').new(base_url, crawled, found, to_follow)
+      Spidr.expects(:site).with(base_url, :proxy => {}).returns
 
       rd.run(to_follow.shift)
     end
@@ -58,13 +58,15 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      search = 'busybox'
+      upstream_credentials_and_search = {
+        search: 'busybox'
+      }
 
       RestClient::Request.expects(:execute)
-        .with({ method: :get, url: base_url.to_s + "v1/search?q=#{search}", headers: {:accept => :json} })
+        .with({ method: :get, url: base_url.to_s + "v1/search?q=busybox", headers: {:accept => :json} })
         .returns({results: ['busybox']}.to_json)
 
-      rd = RepoDiscovery.new(base_url, 'docker', nil, nil, search, crawled, found, to_follow)
+      rd = RepoDiscovery.class_for('docker').new(base_url, crawled, found, to_follow, upstream_credentials_and_search)
       rd.expects(:proxy).returns(nil)
 
       rd.run(to_follow.shift)
@@ -75,13 +77,15 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      search = 'busybox'
+      upstream_credentials_and_search = {
+        search: 'busybox'
+      }
 
       RestClient::Request.expects(:execute)
-        .with({ method: :get, url: base_url.to_s + "v1/search?q=#{search}", proxy: @proxy_url, headers: {:accept => :json} })
+        .with({ method: :get, url: base_url.to_s + "v1/search?q=busybox", proxy: @proxy_url, headers: {:accept => :json} })
         .returns({results: ['busybox']}.to_json)
 
-      rd = RepoDiscovery.new(base_url, 'docker', nil, nil, search, crawled, found, to_follow)
+      rd = RepoDiscovery.class_for('docker').new(base_url, crawled, found, to_follow, upstream_credentials_and_search)
       rd.stubs(:proxy).returns(@http_proxy)
 
       rd.run(to_follow.shift)
@@ -98,17 +102,19 @@ module Katello
       crawled = []
       found = []
       to_follow = [base_url]
-      search = 'busybox'
+      upstream_credentials_and_search = {
+        search: 'busybox'
+      }
 
       RestClient::Request.expects(:execute)
-        .with({ method: :get, url: base_url.to_s + "v1/search?q=#{search}", proxy: @proxy_url, headers: {:accept => :json} })
+        .with({ method: :get, url: base_url.to_s + "v1/search?q=busybox", proxy: @proxy_url, headers: {:accept => :json} })
         .returns({code: Net::HTTPNotFound}.to_json)
 
       RestClient::Request.expects(:execute)
         .with({ method: :get, url: base_url.to_s + "v2/_catalog", proxy: @proxy_url, headers: {:accept => :json} })
         .returns({'repositories' => ['busybox']}.to_json.extend(MockHeaders))
 
-      rd = RepoDiscovery.new(base_url, 'docker', nil, nil, search, crawled, found, to_follow)
+      rd = RepoDiscovery.class_for('docker').new(base_url, crawled, found, to_follow, upstream_credentials_and_search)
       rd.stubs(:proxy).returns(@http_proxy)
 
       rd.run(to_follow.shift)
